Practice/contract_simplifier/requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
blinker==1.9.0
cachetools==6.2.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
cryptography==46.0.1
distro==1.9.0
gitdb==4.0.12
GitPython==3.1.45
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jiter==0.11.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
lxml==6.0.2
MarkupSafe==3.0.2
narwhals==2.5.0
numpy==2.3.3
openai==1.108.2
packaging==25.0
pandas==2.3.2
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.3.0
protobuf==6.32.1
pyarrow==21.0.0
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
pydeck==0.9.1
pypdfium2==4.30.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
pytz==2025.2
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
streamlit==1.49.1
tenacity==9.1.2
toml==0.10.2
tornado==6.5.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0

streamlit
pdfplumber
python-docx
pytesseract
pdf2image
Pillow
openai
